Create rsync_failure_codes option:
User may optionally create a file rsync_failure_codes with a newline (or
space) separated list of rsync return codes that are to be regarded as
complete failure.

If rsync exits with such a code, then the backup will be marked for deletion
during the next ccollect run.

+++ ccollect.sh
@@ -282,7 +282,8 @@ while [ "${i}" -lt "${no_sources}" ]; do
    c_dest="${backup}/destination"
    c_pre_exec="${backup}/pre_exec"
    c_post_exec="${backup}/post_exec"
-   for opt in exclude verbose very_verbose rsync_options summary delete_incomplete remote_host ; do
+   for opt in exclude verbose very_verbose rsync_options summary delete_incomplete \
+         remote_host rsync_failure_codes ; do
       if [ -f "${backup}/$opt" -o -f "${backup}/no_$opt"  ]; then
          eval c_$opt=\"${backup}/$opt\"
       else
@@ -536,16 +537,34 @@ while [ "${i}" -lt "${no_sources}" ]; do
 
    _techo "Transferring files..."
    rsync "$@" "${source}" "${destination_full}"; ret=$?
-
-   #
-   # remove marking here
-   #
-   pcmd rm "${destination_dir}.${c_marker}" || \
-      _exit_err "Removing ${destination_dir}/${c_marker} failed."
-
    _techo "Finished backup (rsync return code: $ret)."
-   if [ "${ret}" -ne 0 ]; then
-      _techo "Warning: rsync exited non-zero, the backup may be broken (see rsync errors)."
+
+   #
+   # Check if rsync exit code indicates failure.
+   #
+   fail=""
+   if [ -f "$c_rsync_failure_codes" ]; then
+      while read line ; do
+         line="${line%% }"
+         if [ -n "$line" ]; then
+            for code in $line ; do
+               [ "$ret" -eq "$code" ] && fail=1
+            done
+         fi
+      done <"$c_rsync_failure_codes"
+   fi
+
+   #
+   # If rsync exit code OK, remove marking.  If exit code non-zero, issue warning.
+   #
+   if [ -n "$fail" ]; then
+      _techo "Warning: rsync failed with return code $ret."
+   else
+      pcmd rm "${destination_dir}.${c_marker}" || \
+         _exit_err "Removing ${destination_dir}/${c_marker} failed."
+      if [ "${ret}" -ne 0 ]; then
+         _techo "Warning: rsync exited non-zero, the backup may be broken (see rsync errors)."
+      fi
    fi
 
    #
